fix(macOS): build with new XCode12 cross-compilation
This is to address onivim/oni2#2550, onivim/oni2#2508, and probably some others that I missed.

This doesn't really get to the "root" of the issue, just ignores the compiler warning for now. It should be safe since we aren't _actually_ cross compiling, but if/when we begin to, we'll need to remove the ignore flag and try to fix the underlying issue.